% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dynRangePlot}
\alias{dynRangePlot}
\title{Produce signal-abundance plot to evaluate dynamic range}
\usage{
dynRangePlot(exDat)
}
\arguments{
\item{exDat}{list, contains input data and stores analysis results}
}
\description{
Produce signal-abundance plot to evaluate dynamic range
}
\examples{
data(SEQC.Example)

exDat <- initDat(datType="count", isNorm=FALSE, exTable=MET.CTL.countDat,
                 filenameRoot="testRun", sample1Name="MET",
                 sample2Name="CTL", erccmix="RatioPair",
                 erccdilution=1/100, spikeVol=1, totalRNAmass=0.500,
                 choseFDR=0.1)

exDat <- est_r_m(exDat)

exDat <- dynRangePlot(exDat)

exDat$Figures$dynRangePlot
}

