\name{ERCCDef}
\alias{ERCCDef}
\docType{data}
\title{
ERCCDef dataframe
}
\description{
ERCC transcript lengths and GC content
}
\format{
  A data frame with 96 observations on the following 3 variables.
  \describe{
    \item{\code{Feature}}{a factor vector}
    \item{\code{Length}}{a numeric vector}
    \item{\code{GC}}{a numeric vector}
  }
}
\details{
Length and GC content of all 96 ERCC controls in NIST SRM 2374
}
\source{
http://tinyurl.com/erccsrm
}
\keyword{datasets}
