\name{ERCCMix1and2}
\alias{ERCCMix1and2}
\docType{data}
\title{
ERCCMix1and2 dataframe
}
\description{
Ambion RatioPair ERCC Mixtures 
}
\format{
  A data frame with 96 observations on the following 4 variables.
  \describe{
    \item{\code{ERCC.AMB.Expected}}{a factor vector of all 96 ERCC control IDs}
    \item{\code{Subpool}}{a factor vector of the ERCC Ratios in each Subpool 
    with levels \code{4:1} \code{1:1} \code{1:1.5} \code{1:2}}
    \item{\code{Mix1Conc.Attomoles_ul}}{a numeric vector of the ERCC concentrations in Mix 1}
    \item{\code{Mix2Conc.Attomoles_ul}}{a numeric vector of the ERCC concentrations in Mix 2}
  }
}
\source{
http://www.lifetechnologies.com/order/catalog/product/4456739
}
\keyword{datasets}
