\name{MET.CTL.countDat}
\alias{MET.CTL.countDat}
\docType{data}
\title{
Rat toxicogenomics count data
}
\description{
RNA-Seq count data from Methimazole and Control rat biological replicates
}
\format{
  A data frame with 16590 observations of the following 7 variables.
  \describe{
    \item{\code{Feature}}{a factor vector of all Endogenous and ERCC transcripts in the experiment}
    \item{\code{MET_1}}{a numeric vector of counts from Methimazole treatment biological replicate 1}
    \item{\code{MET_2}}{a numeric vector of counts from Methimazole treatment biological replicate 2}
    \item{\code{MET_3}}{a numeric vector of counts from Methimazole treatment biological replicate 3}
    \item{\code{CTL_1}}{a numeric vector of counts from Control biological replicate 1}
    \item{\code{CTL_2}}{a numeric vector of counts from Control biological replicate 2}
    \item{\code{CTL_3}}{a numeric vector of counts from Control biological replicate 3}
  }
}

