\name{SEQC.Example}
\alias{SEQC.Example}
\docType{data}
\title{
Example data from SEQC project for erccdashboard analysis
}
\description{
Contains the following 5 itemsL
  MET.CTL.countDat - Rat toxicogenomics count data
MET.CTL.totalReads - Rat toxicogenomics total read data
UHRR.HBRR.arrayDat - UHRR and HBRR Illumina BeadArray data
UHRR.HBRR.countDat - UHRR and HBRR RNA-Seq Illumina count data
UHRR.HBRR.totalReads - UHRR and HBRR sample total read data
}
\usage{data(SEQC.Example)}
\examples{
data(SEQC.Example)
}
\keyword{datasets}
