\name{UHRR.HBRR.arrayDat}
\alias{UHRR.HBRR.arrayDat}
\docType{data}
\title{
UHRR and HBRR Illumina BeadArray data
}
\description{
Unnormalized microarray data from Lab 13 of reference sample interlaboratory study
}
\format{
  A data frame with 17627 observations of the following 7 variables.
  \describe{
    \item{\code{Feature}}{a factor vector of all Endogenous and ERCC transcripts in the experiment}
    \item{\code{UHRR_3}}{a numeric vector of fluorescence intensities from UHRR microarray technical replicate 1}
    \item{\code{UHRR_2}}{a numeric vector of fluorescence intensities from UHRR microarray technical replicate 2}
    \item{\code{UHRR_1}}{a numeric vector of fluorescence intensities from UHRR microarray technical replicate 3}
    \item{\code{HBRR_3}}{a numeric vector of fluorescence intensities from HBRR microarray technical replicate 1}
    \item{\code{HBRR_2}}{a numeric vector of fluorescence intensities from HBRR microarray technical replicate 2}
    \item{\code{HBRR_1}}{a numeric vector of fluorescence intensities from HBRR microarray technical replicate 3}
  }
}
\keyword{datasets}
