\name{UHRR.HBRR.totalReads}
\alias{UHRR.HBRR.totalReads}
\docType{data}
\title{
UHRR and HBRR sample total read data
}
\description{
Total reads per library replicate from FASTQ files
}
\format{
  The format is:
 int [1:8] 138786892 256006510 199468322 431933806 247985592 219383270 251265814 257508210
}
