% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveERCCPlots}
\alias{saveERCCPlots}
\title{Save erccdashboard plots to a pdf file}
\usage{
saveERCCPlots(exDat, plotsPerPg = "manuscript", saveas = "pdf", res = 180,
  plotlist = NULL)
}
\arguments{
\item{exDat}{list, contains input data and stores analysis results}

\item{plotsPerPg}{string, if "manuscript" then the 4 main plots are printed
to one page, if "single" then each plot is printed to page
in the pdf file}

\item{saveas}{Choose file format from "pdf", "jpeg" or "png"}

\item{res}{Choose the file resolution}

\item{plotlist}{list, contains plots to print}
}
\description{
The function savePlots will save selected figures to a pdf file. The default
is the 4 manuscript figures to a single page (plotsPerPg = "manuscript").
If plotsPerPg = "single" then each plot is placed on an
individual page in one pdf file. If plotlist is not defined (plotlist = NULL)
 then all plots in exDat$Figures are printed to the file.
}
\examples{
\donttest{
data(SEQC.Example)

exDat <- initDat(datType="count", isNorm=FALSE, exTable=MET.CTL.countDat,
                 filenameRoot="testRun", sample1Name="MET",
                 sample2Name="CTL", erccmix="RatioPair",
                 erccdilution=1/100, spikeVol=1, totalRNAmass=0.500,
                 choseFDR=0.1)

exDat <- est_r_m(exDat)

exDat <- dynRangePlot(exDat)

exDat <- geneExprTest(exDat)

exDat <- estLODR(exDat, kind="ERCC", prob=0.9)

exDat <- annotLODR(exDat)

#to print 4 plots from manuscript to a single page pdf file
saveERCCPlots(exDat, plotsPerPg = "manuscript")

# or to create a multiple page pdf of all plots produced
saveERCCPlots(exDat, plotsPerPg = "single", plotlist = exDat$Figures)
}
}

